code/requirements.txt
addict==2.1.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-rest-swagger==2.1.2
djangorestframework==3.7.7
drf-openapi==1.3.0
drfdocs==0.0.11
idna==2.6
ipython==1.2.1
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
Pygments==2.2.0
PyMySQL==0.8.0
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
uritemplate==3.0.0
urllib3==1.22
